+++ client/src/components/ColorPage.js
@@ -1,14 +1,22 @@
 import React from 'react';
 import { connect } from 'react-redux';
+import { ToastContainer, toast, Zoom } from 'react-toastify';
 import PropTypes from 'prop-types';
 
 import { getColors } from '../actions/colorActions';
 
 import ColorCard from './ColorCard';
+import NewColorCard from './NewColorCard';
 import Loader from './Loader';
 import NotFound from './NotFound';
 
 class ColorPage extends React.Component {
+	constructor(props) {
+		super(props);
+
+		this.create = this.create.bind(this);
+	}
+
 	componentDidMount() {
 		const { type } = this.props.match.params;
 		if (type === 'swatches' || type === 'palettes' || type === 'gradients') {
@@ -23,9 +31,25 @@ class ColorPage extends React.Component {
 		}
 	}
 
+	create() {
+		if (this.props.user.authenticated) {
+			this.props.history.push('/profile/me');
+		} else {
+			toast.info('You need to login to create new color', {
+				position: 'bottom-right',
+				autoClose: 2000,
+				hideProgressBar: true,
+				closeOnClick: true,
+				pauseOnHover: false,
+				className: 'toast'
+			});
+		}
+	}
+
 	render() {
 		const { type } = this.props.match.params;
 		const { loading } = this.props.color;
+		const colorType = type === 'swatches' ? 'swatch' : type.substring(0, type.length - 1);
 
 		return (
 			<main className="colorpage">
@@ -42,6 +66,9 @@ class ColorPage extends React.Component {
 									{this.props.color[`${type}`].map((color, index) => (
 										<ColorCard key={color._id} color={color} index={index} />
 									))}
+									<NewColorCard onClick={this.create} index={this.props.color[`${type}`].length}>
+										Create New {colorType}
+									</NewColorCard>
 								</div>
 							)
 						) : (
@@ -51,17 +78,20 @@ class ColorPage extends React.Component {
 						<NotFound />
 					)}
 				</div>
+				<ToastContainer transition={Zoom} />
 			</main>
 		);
 	}
 }
 
-const mapStateToProps = ({ color }) => ({ color });
+const mapStateToProps = ({ user, color }) => ({ user, color });
 
 ColorPage.propTypes = {
 	match: PropTypes.object.isRequired,
 	location: PropTypes.object.isRequired,
+	history: PropTypes.object.isRequired,
 	color: PropTypes.object.isRequired,
+	user: PropTypes.object.isRequired,
 	getColors: PropTypes.func.isRequired
 };